[PBIOS-460] Fix user kit name fonts (#413)
**What does this PR do?** A clear and concise description with your
runway ticket url.
We recently changed the User kit to make all the name font bold. The
problem is that Connect uses the light variant for the name.

OBS: this story fix this problem allowing the user to pass a custom font
for the kit, but we need to discuss with the team if we really want this
change or not.



**How to test?** Steps to confirm the desired behavior:
1. Go to '...'
2. Click on '....'
3. Scroll down to '....'
4. See addition/change

### Checklist
- [ ] **LABELS** - Add a label: `breaking`, `bug`, `improvement`,
`documentation`, or `enhancement`. See
[Labels](https://github.com/powerhome/playbook-apple/labels) for
descriptions.
- [ ] **RELEASES** - Add the appropriate label: `Ready for Testing` /
`Ready for Release`
- [ ] **TESTING** - Have you tested your story?