[RANGR-966] Clean Keychain on Failed Jobs (#520)
Task on Runway
[RANGR-966](https://nitro.powerhrg.com/runway/backlog_items/RANGR-966).

Context on this work:

Yesterday
[worker10](https://ci.powerapp.cloud/blue/organizations/jenkins/powerhome%2Fconnect-apple/detail/PR-1333/9/pipeline/)
and today
[worker11](https://ci.powerapp.cloud/blue/organizations/jenkins/powerhome%2Fconnect-apple/detail/PR-1340/5/pipeline/137/)
started to fail on the archive step.

I fixed it, but I will try to break it down as best as I can, so if this
happens again there are more devs enabled to fix it:

- The error was: error: Provisioning profile “Connect v3 Beta Siri
Extension” doesn’t include signing certificate “iPhone Distribution:
Power Home Remodeling Group, LLC”. (in target ‘Connect Siri Extension
iOS’ from project ‘Connect’) (and every other profile.

- Yesterday I merged to iOS-provisioning-profiles the new certificate
that will be used to, for now, regenerate the Beta and RC profiles, and
in the future the production ones as well.

- We can have two iOS Distribution profiles at the same time, but Apple
generates both with the following name “iPhone Distribution: Power Home
Remodeling Group, LLC”, which we can’t change.

- As the certificate I committed is newer, Xcode tries to use it instead
of the older one, and as we don’t have control over the naming, we can’t
change that. That is why, on the CI, on the keychain.sh file, we create
a custom keychain only for that specific build, and in this step, we can
specify which .p12 file we will use.

- The current certificate we use is called ios_distribution.p12, and the
new one I created is called
ios_distribution_certificateID_expirationDate.p12. Nitro Mobile
Components and Playbook Swift are already using the second one.

So, if we have specific custom keychains for each build, why did it
fail?

- Playbook Swift was only cleaning the derived data and custom keychain
when a build was successful, it was not cleaning when the build failed.
As a retry won’t guarantee that a job will run on the same machine,
Playbook left some keychains on these machines using the new
certificate, and this was causing the connect-apple builds to fail.

- This PR fix Playbook CI to always delete the keychain, as the other
CIs do.

- If for some reason any project fails to delete it and we see this bug
again, we just need to access the machine -> Keychain Access -> Delete
any old custom keychain still there.

- This won’t happen as soon as all projects are migrated to the new
certificate, but it can happen again in some years when we need to
update the certs again.