[RANGR-1025] Enable Kits to Allow Inactive User (#534)
**What does this PR do?** 

[RANGR-1025] Enable Kits to Allow Inactive User

https://runway.powerhrg.com/backlog_items/RANGR-1025

**Notes:** 

Updated Components:

PBUser -> Name font & Avatar. Removed hasInactivebadge and just used
isActive to avoid redundant code.
PBMultipleUserStacked -> Added inactive but because we are using mock
data and not fetching it, they will all be inactive in Showcase.
PBMultipleUsers -> Added inactive but because we are using mock data and
not fetching it, they will all be inactive in Showcase.
PBMessage -> Name font & Avatar. This uses anyview for an avatar so
couldn’t enable inactive avatar variant but its ok because it can easily
be handled in connect.


### Checklist
- [x] **LABELS** - Add a label: `breaking`, `bug`, `improvement`,
`documentation`, or `enhancement`. See
[Labels](https://github.com/powerhome/playbook-apple/labels) for
descriptions.
- [x] **RELEASES** - Add the appropriate label: `Ready for Testing` /
`Ready for Release`
- [x] **TESTING** - Have you tested your story?